1.struct d_int changed from int to uint8_t  2. dessert_search_con changed return value 0 win  1 is fail
git-svn-id: https://svn.imp.fu-berlin.de/des-testbed/Software/DES-SERT/libdessert/trunk@7370 ec5dfc5b-5239-0410-a031-e37cc464ff32

+++ include/dessert.h
@@ -576,8 +576,8 @@ void _dessert_log(int level, const char* func, const char* file, int line, const
 ******************************************************************************/
 
 struct d_int{
-  int value;
-  int number;
+  uint8_t value;
+  uint8_t number;
 };
 
 /******************************************************************************
